Addded new sparql checks
Exclude objects that have two parents from result. This way, if an
object was marked for deletion, but has a relationship to a Collection,
then don't remove it!. This is important because the batch could get
into islanders basic functionality and remove a lot of objects in it's
way. Also adds XACML check on sparql.